Create README.md
Description of what it is. Their course basically covered the same algorithm things that I worked on, but all in one course instead of spread out over a lot of courses. The truth of the matter is that I haven't written custom data structures in my career except for a handful of times. I just use the built-in ones, and most of the applications I've worked on don't have anything complex like that.